requirements.txt
streamlit>=1.35,<1.40
streamlit-drawable-canvas>=0.9.3
streamlit-shortcuts>=1.2.0
ultralytics>=8.1.0
torch
pillow
numpy
pandas
opencv-python-headless
